backend/requirements.txt
# FastAPI & ASGI Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Upload
python-multipart==0.0.12

# Data Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
motor==3.6.0

# AI & ML
google-generativeai==0.8.3

# Firebase
firebase-admin==6.5.0

# Environment Variables
python-dotenv==1.0.1

# Image Processing
Pillow==10.4.0

# HTTP Client (for image search API)
httpx==0.27.2
